backend/pyproject.toml
[project]
name = "rag-agent-framework"
version = "0.1.0"
description = "Multi-agent RAG framework with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.12"

authors = [
    { name = "Joab David Johanan" }
]

dependencies = [

    # FastAPI stack
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",

    # LangChain ecosystem
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-groq>=0.1.0",
    "langgraph>=0.2.0",

    # Redis cache
    "redis>=5.0.0",

    # Vector databases
    "faiss-cpu>=1.8.0",
    "pinecone-client>=4.1.0",

    # Embeddings / models
    "sentence-transformers>=2.7.0",

    # Utilities
    "numpy>=1.26.0",
    "tqdm>=4.66.0",

    # Web search
    "langchain-tavily>=0.1.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]